@@ -1,6 +1,6 @@
 {
-    "name": "eigopost-client",
-    "version": "1.0.0",
+    "name": "eigopost - client",
+    "version": "2.0.0",
     "license": "MIT",
     "main": "index.js",
     "scripts": {
@@ -9,17 +9,17 @@
         "start": "node server.js"
     },
     "dependencies": {
-        "axios": "^0.21.3",
-        "bootstrap": "^5.1.2",
-        "formik": "^2.1.4",
+        "axios": "^0.24.0",
+        "bootstrap": "^5.1.3",
+        "formik": "^2.2.9",
         "history": "^4.10.1",
         "moment": "^2.29.1",
         "path": "^0.12.7",
         "prop-types": "^15.7.2",
         "query-string": "^6.11.0",
-        "react": "^16.8.6",
-        "react-dom": "^16.8.6",
-        "react-router-dom": "^5.2.0",
+        "react": "^17.0.2",
+        "react-dom": "^17.0.2",
+        "react-router-dom": "^6.2.1",
         "react-spring": "^8.0.27",
         "rxjs": "^6.3.3",
         "yup": "^0.28.1"
@@ -31,15 +31,15 @@
         "@babel/preset-env": "^7.12.1",
         "@babel/preset-react": "^7.0.0",
         "babel-loader": "^8.1.0",
-        "css-loader": "^6.2.0",
+        "css-loader": "^6.5.1",
         "file-loader": "^6.2.0",
         "html-webpack-plugin": "^5.3.2",
         "sass": "^1.42.1",
         "sass-loader": "^12.1.0",
         "style-loader": "^3.2.1",
         "svg-url-loader": "^7.1.1",
-        "webpack": "^5.52.0",
-        "webpack-cli": "^4.8.0",
-        "webpack-dev-server": "^4.1.0"
+        "webpack": "^5.65.0",
+        "webpack-cli": "^4.9.1",
+        "webpack-dev-server": "^4.6.0"
     }
 }
