@@ -1,18 +1,16 @@
 import React from 'react';
-import { Route, Switch } from 'react-router-dom';
+import { Route, Routes } from 'react-router-dom';
 
 import { List } from './List';
 import { AddEdit } from './AddEdit';
 
-function Users({ match }) {
-    const { path } = match;
-    
+function Users() {
     return (
-        <Switch>
-            <Route exact path={path} component={List} />
-            <Route path={`${path}/add`} component={AddEdit} />
-            <Route path={`${path}/edit/:id`} component={AddEdit} />
-        </Switch>
+        <Routes>
+            <Route path='/' element={<List/>} />
+            <Route path='add' element={<AddEdit/>} />
+            <Route path='edit/:id' element={<AddEdit/>} />
+        </Routes>
     );
 }
 
