@@ -1,50 +1,51 @@
 import React, {useState, useEffect} from 'react';
+import {Link, Route, Routes} from 'react-router-dom';
 import {accountService} from '@/_services';
-import {Link, Route, Switch, Redirect, useLocation} from 'react-router-dom';
-import {Nav, Alert, NotFound, Footer, PrivateRoute, Unauthorized} from '@/_components';
-import {Role} from '@/_helpers';
+import {Nav, Footer} from '@/layout';
+import {keepTheme, Alerts, NotFound, Unauthorized} from '@/_components';
+import {Role, RouteAdmin, RouteUser} from '@/_helpers';
 import {Landing, About, Contact, Terms, Privacy} from '@/home';
 import {Profile} from '@/profile';
 import {Admin} from '@/admin';
 import {Account} from '@/account';
 import Posts from '../posts/Posts';
 import Post from '../posts/Post';
-import FormPost from '../posts/FormPost';
-import {keepTheme} from '../_components/ThemeToggle';
+import PostForm from '../posts/PostForm';
 
 const App = () => {
-  const { pathname } = useLocation();
   const [user, setUser] = useState({});
   useEffect(() => {
-      const subscription = accountService.user.subscribe(x => setUser(x));
-      return subscription.unsubscribe;
+    const subscription = accountService.user.subscribe(x => setUser(x));
+    return subscription.unsubscribe;
   }, []);
   useEffect(() => {
-      keepTheme();
+    keepTheme();
   })
-
   return (
     <div className="container-fluid container-main px-0">
       <Nav />
-      <Alert />
+      <Alerts />
       <div className="global-mh">
-        <Switch>
-          <Redirect from="/:url*(/+)" to={pathname.slice(0, -1)} />
-          <Route exact path="/" component={Landing} />
-          <Route path="/about" component={About} />
-          <Route path="/contact" component={Contact} />
-          <Route path="/terms-of-use" component={Terms} />
-          <Route path="/privacy-policy" component={Privacy} />
-          <Route path="/profile" component={Profile} />
-          <PrivateRoute path="/admin" roles={[Role.Admin]} component={Admin} />
-          <Route path="/account" component={Account} />
-          <Route path="/new-post" component={FormPost} />
-          <Route exact path="/posts" component={Posts} />
-          <Route exact path="/posts/:slug" component={Post} />
-          <Route path="/not-found" component={NotFound} />
-          <Route path="/unauthorized" component={Unauthorized} />
-          <Route component={NotFound} />
-        </Switch>
+        <Routes>
+          <Route path="/" element={<Landing />} />
+          <Route path="about" element={<About />} />
+          <Route path="contact" element={<Contact />} />
+          <Route path="terms-of-use" element={<Terms />} />
+          <Route path="privacy-policy" element={<Privacy />} />
+          <Route path="profile/*" element={<Profile />} />
+          <Route path="admin/*" element={
+             <RouteAdmin roles={[Role.Admin]}>
+               <Admin />
+             </RouteAdmin>
+          }/>
+          <Route path="account/*" element={<Account />} />
+          <Route path="new-post" element={ <PostForm />}/>
+          <Route path="posts" element={<Posts />} />
+          <Route path="posts/:slug" element={<Post />} />
+          <Route path="not-found" element={<NotFound />} />
+          <Route path="unauthorized" element={<Unauthorized />} />
+          <Route path="*" element={<NotFound />} />
+        </Routes>
       </div>
       <Footer/>
     </div>
