@@ -2,6 +2,7 @@ import React from "react";
 import {NavLink} from 'react-router-dom';
 import { useTransition, animated } from "react-spring";
 import close from "../assets/icon-close.svg";
+import './Modal.css';
 
 const ModalDelete = (props) => {
   const transitions = useTransition(props.modalVis, null, {
@@ -16,6 +17,7 @@ const ModalDelete = (props) => {
             style={style}
             closeModal={() => props.setModalVisOff(false)}
             deleteModal={() => props.deleteAccount()}
+            isSubmitting={props.isSubmitting}
             key={key}
           />
         )
@@ -24,7 +26,11 @@ const ModalDelete = (props) => {
   );
 }
 
-const ModalDeleteSub = ({ style, deleteModal, closeModal }) => (
+const ModalDeleteSub = ({ style,
+                          deleteModal,
+                          closeModal,
+                          isSubmitting
+                       }) => (
   <animated.div style={style}
                 className="ep-modal">
     <div className="ep-modal-header">
@@ -33,14 +39,16 @@ const ModalDeleteSub = ({ style, deleteModal, closeModal }) => (
       </h3>
     </div>
     <div className="ep-modal-body">
-      <h5></h5>
+      <h5>We will miss you...</h5>
     </div>
     <div className="ep-modal-footer">
-      <button className="btn btn-primary me-2"
+      <button type="button"
+              className="btn btn-primary me-2"
               onClick={deleteModal}>
-        delete
+              {isSubmitting ? <span className="ep-spinner ep-spinner-sm mr-1"></span> : "delete"}
       </button>
-      <button className="btn btn-secondary ps-2"
+      <button type="button"
+              className="btn btn-secondary ps-2"
               onClick={closeModal}>
         go back
       </button>
