@@ -1,15 +1,15 @@
 import React from "react";
 import {NavLink} from 'react-router-dom';
-import {useFormikContext} from 'formik';
 import {useTransition, animated} from "react-spring";
 import close from "../assets/icon-close.svg";
+import './Modal.css';
+
 
 const ModalSend = (props) => {
-  const {values} = useFormikContext();
   const transitions = useTransition(props.modalVis, null, {
     from: { opacity: 0, transform: "translateY(-2.5rem)" },
-    enter: { opacity: 1, transform: "translateY(0rem)" },
-    leave: { opacity: 0, transform: "translateY(-2.5rem)" }
+    enter:{ opacity: 1, transform: "translateY(0rem)" },
+    leave:{ opacity: 0, transform: "translateY(-2.5rem)" }
   });
   return (
     <div className="container-fluid">
@@ -17,8 +17,9 @@ const ModalSend = (props) => {
           <ModalSendSub
             style={style}
             closeModal={() => props.setModalVisOff(false)}
-            sendData={() => props.send(values)}
+            sendData={() => props.send(props.values, props.setSubmitting)}
             statement={props.statement}
+            isSubmitting={props.isSubmitting}
             key={key}
           />
       ))}
@@ -26,7 +27,12 @@ const ModalSend = (props) => {
   );
 }
 
-const ModalSendSub = ({ style, sendData, closeModal, statement }) => (
+const ModalSendSub = ({ style,
+                        sendData,
+                        closeModal,
+                        statement,
+                        isSubmitting
+                      }) => (
   <animated.div style={style}
                 className="ep-modal">
     <div className="ep-modal-header">
@@ -38,11 +44,15 @@ const ModalSendSub = ({ style, sendData, closeModal, statement }) => (
       <h5></h5>
     </div>
     <div className="ep-modal-footer">
-      <button type="button" className="btn btn-primary me-2"
+      <button type="button"
+              className="btn btn-primary me-2"
               onClick={sendData}>
-        yes, post
+              {isSubmitting &&
+              <span className="ep-spinner ep-spinner-sm mr-1"></span>}
+        Post
       </button>
-      <button type="button" className="btn btn-secondary ps-2"
+      <button type="button"
+              className="btn btn-secondary ps-2"
               onClick={closeModal}>
         go back
       </button>
