@@ -3,9 +3,9 @@ import {useTransition, animated} from "react-spring";
 
 const PopoverVote = (props) => {
   const transitions = useTransition(props.popoverOpen, null, {
-     from: { opacity: 0, transform: "translateY(-.188rem)"},
+     from: { opacity: 0, transform: "translateY(-0.188rem)"},
     enter: { opacity: 1, transform: "translateY(0rem)"},
-    leave: { opacity: 0, transform: "translateY(.188rem)"}
+    leave: { opacity: 0, transform: "translateY(0.188rem)"}
   });
   return (
     <div className="container">
