@@ -4,12 +4,14 @@ import notFoundLogo from '../assets/img-logo-404.png';
 const Unauthorized = () => {
   document.title = "Eigopost | Unauthorized"
   return (
-    <div className="container unauthorized text-center fade-in">
-        <img src={notFoundLogo}
-             className="logo-404 fade-in"
-             alt="logo"
-        />
-        <h3>You do not have permission to access that page.</h3>
+    <div className="container unauthorized text-center">
+      <img src={notFoundLogo}
+           className="logo-404 fade-in"
+           alt="unauthorized-logo"
+      />
+      <h3>
+      This page is only accessible to administrators.
+      </h3>
     </div>
   );
 }
