@@ -1,20 +1,12 @@
-export * from './Alert';
-export * from './Comment';
+export * from './Alerts';
 export * from './Error';
-export * from './Footer';
 export * from './Loader';
-export * from './LoaderTimeout';
 export * from './ModalAccounts';
 export * from './ModalDelete';
 export * from './ModalSend';
-export * from './Nav';
-export * from './NavMobile';
 export * from './NotFound';
 export * from './PopoverProfile';
 export * from './PopoverVote';
-export * from './PrivateRoute';
-export * from './SkeletonEffects';
-export * from './SkeletonPosts';
-export * from './SkeletonPostsTimeout';
+export * from './Skeleton';
 export * from './ThemeToggle';
 export * from './Unauthorized';
