@@ -1,48 +1,48 @@
-import React, { useState, useEffect } from 'react';
-import { Link } from 'react-router-dom';
+import React, {useState, useEffect} from 'react';
+import {Link, useNavigate, useLocation} from 'react-router-dom';
 import queryString from 'query-string';
+import {accountService, alertService} from '@/_services';
 
-import { accountService, alertService } from '@/_services';
+function VerifyEmail() {
+  const navigate = useNavigate();
+  const pathname = useLocation();
+  console.log("PATHNALE", pathname);
+  const EmailStatus = {
+    Verifying: 'Verifying',
+    Failed: 'Failed'
+  }
+  const [emailStatus, setEmailStatus] = useState(EmailStatus.Verifying);
+  useEffect(() => {
+    const { token } = queryString.parse(location.search);
+    // remove token from url to prevent http referer leakage
+    navigate(location.pathname);
+    accountService.verifyEmail(token)
+      .then(() => {
+        alertService.success('Verification successful, you can now login',
+         { keepAfterRouteChange: true });
+        navigate('/account/login');
+      })
+      .catch(() => {
+        setEmailStatus(EmailStatus.Failed);
+      });
+  }, []);
 
-function VerifyEmail({ history }) {
-    const EmailStatus = {
-        Verifying: 'Verifying',
-        Failed: 'Failed'
+  function getBody() {
+    switch (emailStatus) {
+      case EmailStatus.Verifying:
+        return <div>Verifying...</div>;
+      case EmailStatus.Failed:
+        return <div>Verification failed, you can also verify your account using the
+         <Link to="/account/forgot-password">
+           forgot password
+         </Link> page.</div>;
     }
-
-    const [emailStatus, setEmailStatus] = useState(EmailStatus.Verifying);
-
-    useEffect(() => {
-        const { token } = queryString.parse(location.search);
-
-        // remove token from url to prevent http referer leakage
-        history.replace(location.pathname);
-
-        accountService.verifyEmail(token)
-            .then(() => {
-                alertService.success('Verification successful, you can now login', { keepAfterRouteChange: true });
-                history.push('login');
-            })
-            .catch(() => {
-                setEmailStatus(EmailStatus.Failed);
-            });
-    }, []);
-
-    function getBody() {
-        switch (emailStatus) {
-            case EmailStatus.Verifying:
-                return <div>Verifying...</div>;
-            case EmailStatus.Failed:
-                return <div>Verification failed, you can also verify your account using the <Link to="forgot-password">forgot password</Link> page.</div>;
-        }
-    }
-
-    return (
-        <div>
-            <h3 className="card-header">Verify Email</h3>
-            <div className="card-body">{getBody()}</div>
-        </div>
-    )
+  }
+  return (
+    <div>
+      <h3 className="card-header">Verify Email</h3>
+      <div className="card-body">{getBody()}</div>
+    </div>
+  )
 }
-
-export { VerifyEmail };
+export {VerifyEmail};
