@@ -1,6 +1,6 @@
+export * from './Account';
 export * from './ForgotPassword';
 export * from './Login';
 export * from './Register';
 export * from './ResetPassword';
-export * from './Switch';
 export * from './VerifyEmail';
