@@ -1,19 +1,22 @@
 import React from 'react';
-import { Link } from 'react-router-dom';
-import Colors from '../_helpers/colors';
+import {Link} from 'react-router-dom';
+import Colors from './colors';
+import './Admin.css';
 
-function Overview({ match }) {
-    const { path } = match;
-
-    return (
-        <div className="">
-            <h1 className="">Admin</h1>
-            <p className="">This section can only be accessed by administrators.</p>
-            <p><Link to={`${path}/users`}>Manage Users</Link></p>
-            <br />
-            <Colors />
-        </div>
-    );
+function Overview() {
+  return (
+    <div className="">
+      <h1 className="">Admin</h1>
+        <Link to='users'
+              className="btn btn-primary"
+              style={{marginBottom:"1rem"}}
+        >
+          Manage Users
+        </Link>
+      <br />
+      <p>Site Color Scheme</p>
+      <Colors />
+    </div>
+  );
 }
-
-export { Overview };
+export {Overview};
