@@ -1,5 +1,5 @@
 export * from './About';
 export * from './Contact';
 export * from './Landing';
-export * from './Terms';
 export * from './Privacy';
+export * from './Terms';
